+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Update APE dependency to 1.1.9.
 - Change "CWL (beta)" download to "Abstract CWL" download.
+- Make it possible to configure the size limit of file uploads in the .env files (default: 10MB).
 
 ### Fixed